Relocated functions from PortableContact to GServer
